refactor: 사용자 입력 대본의 model과 script를 전역으로 관리.
- 사용자 입력 대본을 다른 feature에서도 사용하게 됨에 따라 사용자 대본의 model과 대본 상태 관리 controller를 common폴더에서 전역으로 관리

#TT-88 scriptInput
