+++ lib/features/common/controllers/before_recode_script.dart
@@ -1,0 +1,13 @@
+import 'package:get/get.dart';
+
+class BeforeRecodeScriptCtr extends GetxController{
+
+  String? _script;
+
+  String? get script => _script;
+
+  void updateScript(String? newScript) {
+    _script = newScript;
+  }
+
+}
