+++ lib/features/scriptInput/viewModel/script_input_controller.dart
@@ -1,25 +1,29 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
+import 'package:swm_peech_flutter/features/common/controllers/before_recode_script.dart';
 import 'package:swm_peech_flutter/features/scriptInput/dataSource/script_expected_time_data_source.dart';
 
-import '../model/script_expected_time_model.dart';
+import '../../common/models/script_expected_time_model.dart';
+
 
 
 class ScriptInputController extends GetxController {
+
+  //녹음 연습 전 스크립트 컨트롤러
+  final BeforeRecodeScriptCtr _beforeRecodeScriptCtr = BeforeRecodeScriptCtr();
 
   //대본 예상 시간 데이터 소스
   final ScriptExpectedTimeDataSource _scriptExpectedTimeDataSource = ScriptExpectedTimeDataSource();
 
   //대본 입력 데이터
-  String? _script;
   Rx<String?> script = Rx<String?>(null);
 
   //발표 전 대본 기반 에상 시간
   ScriptExpectedTimeModel? _scriptExpectedTime;
   Rx<ScriptExpectedTimeModel?> scriptExpectedTime = Rx<ScriptExpectedTimeModel?>(null);
 
+  //스크립트 예상시간 로딩 유무
   Rx<bool> isLoading = false.obs;
-
 
   //대본 입력 TextField 텍스트 컨트롤러
   TextEditingController textEditingController = TextEditingController();
@@ -40,8 +44,8 @@ class ScriptInputController extends GetxController {
   }
 
   void updateScript(String newScript) {
-    _script = newScript;
-    script.value = _script;
+    _beforeRecodeScriptCtr.updateScript(newScript);
+    script.value = _beforeRecodeScriptCtr.script;
   }
 
   void getExpectedTime() async {
